capstone_1/requirements.txt
flask
tensorflow==2.17.0
numpy>=1.23.5
Pillow
waitress